allow subfiles for nft creation
This commit not only simplifies the interface for the nft creation method, it also enhances it by allowing to pass additional files that are used within the `files` section in the metadata.

This commit also introduces a simple class to store file information that can be used to pass a preview image and n additional files.

The API is a bit strange in the sense that all file objects can carry variables. It's unclear to me what happens if they overwrite each other. Also, since the variables are passed with either `previewImageNft` or the subfiles objects I would expect them to only be applied within that scope, but apparently that's not the case: At least the variables passed with `previewImageNft` are also applied outside this scope.